Move routes to own file. Extract plugin ID to const/file.
