+++ x-pack/plugins/integrations_manager/public/register_feature.ts
@@ -8,14 +8,15 @@ import {
   FeatureCatalogueCategory,
   FeatureCatalogueRegistryProvider,
 } from 'ui/registry/feature_catalogue';
+import { ID } from '../common/constants';
 
 // This defines what shows up in the registry found at /app/kibana#/home and /app/kibana#/home/feature_directory
 FeatureCatalogueRegistryProvider.register(() => ({
-  id: 'integrations_manager',
+  id: ID,
   title: 'Integrations Manager',
   description: 'Install and manage your elastic data ingest integrations',
   icon: 'merge',
-  path: `/app/integrations_manager`,
+  path: `/app/${ID}`,
   showOnHomePage: true,
   category: FeatureCatalogueCategory.DATA,
 }));
